# contracts/openapi.yml
openapi: 3.0.3
info:
  title: Sentiment Search API
  version: 2.0.0
  description: >
    Three-stage pipeline: (1) stream stock info, (2) stream articles per ticker,
    (3) stream per-article sentiment scores.

paths:
  /api/tickers:
    get:
      summary: Stream stock records as NDJSON. Accepts either a search query or a list of ticker IDs.
      parameters:
        - in: query
          name: q
          schema:
            type: string
          description: Search query (company name, sector). Mutually exclusive with tickerIds.
        - in: query
          name: tickerIds
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
          description: Repeat the param (?tickerIds=AAPL&tickerIds=MSFT). Mutually exclusive with q.
      responses:
        "200":
          description: NDJSON stream of Stock records (one Stock per line, or a SearchError line).
          content:
            application/x-ndjson:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Stock"
                  - $ref: "#/components/schemas/SearchError"
        "400":
          description: Missing or invalid query parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"
      tags:
        - webclient-out
        - core-in

  /api/tickers/articles:
    get:
      summary: Stream articles for a list of tickers as NDJSON.
      parameters:
        - in: query
          name: tickerIds
          required: true
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
          description: Repeat the param (?tickerIds=AAPL&tickerIds=MSFT).
      responses:
        "200":
          description: NDJSON stream — one TickerArticles per ticker (or a SearchError line).
          content:
            application/x-ndjson:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/TickerArticles"
                  - $ref: "#/components/schemas/SearchError"
        "400":
          description: Missing or invalid query parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"
      tags:
        - webclient-out
        - core-in

  /api/tickers/trending:
    get:
      summary: Stream the current trending US tickers as NDJSON
      responses:
        "200":
          description: NDJSON stream of Stock records (one Stock per line, or a SearchError line).
          content:
            application/x-ndjson:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Stock"
                  - $ref: "#/components/schemas/SearchError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"
      tags:
        - webclient-out
        - core-in

  /api/tickers/{tickerId}/articles:
    get:
      summary: Stream articles for a single ticker as NDJSON. Optionally pin to one or more event timestamps.
      parameters:
        - in: path
          name: tickerId
          required: true
          schema:
            type: string
          description: Ticker symbol (e.g. AAPL)
        - in: query
          name: eventTSec
          schema:
            type: array
            items:
              type: integer
          style: form
          explode: true
          description: Optional. Per-event timestamps (Unix seconds). Repeat the param. Each produces one TickerArticles line.
        - in: query
          name: intervalSec
          schema:
            type: integer
            minimum: 1
          description: Optional. Article-window size in seconds (fromSec = eventTSec - intervalSec, or now - intervalSec for no-event call).
      responses:
        "200":
          description: NDJSON stream — one TickerArticles per event (or a single result when no events given).
          content:
            application/x-ndjson:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/TickerArticles"
                  - $ref: "#/components/schemas/SearchError"
        "400":
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"
        "503":
          description: Article data unavailable (upstream failure)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"
      tags:
        - webclient-out
        - core-in

  /api/tickers/{tickerId}/articles/sentiment:
    get:
      summary: Stream per-article sentiment scores as NDJSON.
      parameters:
        - in: path
          name: tickerId
          required: true
          schema:
            type: string
          description: Ticker symbol (e.g. AAPL)
        - in: query
          name: articleUrl
          required: true
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
          description: Repeat the param (?articleUrl=https://...&articleUrl=https://...). One SourceResult emitted per scored article.
      responses:
        "200":
          description: NDJSON stream — one SourceResult per line as each article is scored (cached ones emit immediately).
          content:
            application/x-ndjson:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/SourceResult"
                  - $ref: "#/components/schemas/SearchError"
        "400":
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"
        "503":
          description: Sentiment data unavailable (upstream failure)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"
      tags:
        - webclient-out
        - core-in

  /api/tickers/{tickerId}/candles:
    get:
      summary: Get OHLC price candles for a ticker
      parameters:
        - in: path
          name: tickerId
          required: true
          schema:
            type: string
          description: Ticker symbol (e.g. AAPL)
        - in: query
          name: duration
          required: true
          schema:
            type: string
            enum: [1D, 1W, 1M, 1Y, today]
          description: Lookback window. `today` starts at UTC midnight; others go back from now.
        - in: query
          name: interval
          required: true
          schema:
            type: string
            enum: [5m, 30m, 1d]
          description: Candle interval.
      responses:
        "200":
          description: OHLC candle series
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandleSeries"
        "400":
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"
        "404":
          description: No price data for this ticker/window
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"
        "503":
          description: Price data unavailable (upstream failure)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"
      tags:
        - webclient-out
        - core-in

  /api/tickers/{tickerId}/peers:
    get:
      summary: Stream peer companies for a ticker as NDJSON
      parameters:
        - in: path
          name: tickerId
          required: true
          schema:
            type: string
          description: Ticker symbol to find peers for (e.g. AAPL)
      responses:
        "200":
          description: NDJSON stream of peer Stock records (one per line, cached peers emitted first).
          content:
            application/x-ndjson:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Stock"
                  - $ref: "#/components/schemas/SearchError"
        "503":
          description: Peer lookup unavailable (upstream failure)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"
      tags:
        - webclient-out
        - core-in

  /api/auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "200":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Username already taken or missing credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"
      tags:
        - webclient-out
        - core-in

  /api/auth/login:
    post:
      summary: Login to an existing user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"
      tags:
        - webclient-out
        - core-in

  /api/lists:
    get:
      summary: Get all user lists with their items
      responses:
        "200":
          description: The user's lists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/List"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"
      tags:
        - webclient-out
        - core-in
    post:
      summary: Create a new list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListRequest"
      responses:
        "200":
          description: Created list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/List"
        "400":
          description: Missing name field
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"
      tags:
        - webclient-out
        - core-in

  /api/lists/{id}:
    patch:
      summary: Rename a list
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenameListRequest"
      responses:
        "200":
          description: Renamed list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListSummary"
        "400":
          description: Missing name field
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"
        "404":
          description: List not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"
      tags:
        - webclient-out
        - core-in
    delete:
      summary: Delete a list
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"
      tags:
        - webclient-out
        - core-in

  /api/lists/{id}/items:
    post:
      summary: Add a ticker to a list
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddListItemRequest"
      responses:
        "200":
          description: Added
        "400":
          description: Missing ticker or ticker not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"
      tags:
        - webclient-out
        - core-in

  /api/lists/{id}/items/{ticker}:
    delete:
      summary: Remove a ticker from a list
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: path
          name: ticker
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Removed
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"
      tags:
        - webclient-out
        - core-in

  /api/lists/stream:
    get:
      summary: Stream per-article sentiment updates for user's watchlist tickers (SSE)
      description: >
        Server-Sent Events stream. Each `data:` line is a JSON-encoded
        `SentimentStreamEvent`. On each scored article for a watched ticker the
        server emits a `SOURCE_UPDATE` event containing the article's SourceResult
        and the recomputed ticker avgScore. Pass the JWT via the `token` query
        parameter when using the browser `EventSource` API (which cannot set
        custom headers).
      parameters:
        - in: query
          name: token
          schema:
            type: string
          description: JWT token (alternative to Authorization header, required for EventSource)
      responses:
        "200":
          description: SSE stream of sentiment update events (use EventSource, not fetch)
          content:
            text/event-stream:
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"
      tags:
        - webclient-out
        - core-in

components:
  schemas:
    Stock:
      $ref: "./schemas/Stock.json"

    TickerArticles:
      $ref: "./schemas/TickerArticles.json"

    SourceResult:
      $ref: "./schemas/SourceResult.json"

    Candle:
      $ref: "./schemas/Candle.json"

    SourceSchema:
      $ref: "./schemas/Source.json"

    HttpError:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message safe to display to the user.
        code:
          type: string
          description: Machine-readable error code (e.g. UPSTREAM_UNAVAILABLE, USERNAME_TAKEN).
      required:
        - error
        - code

    SearchError:
      type: object
      description: An error chunk emitted mid-stream for a specific ticker (or the whole request).
      properties:
        error:
          type: string
        code:
          type: string
        ticker:
          type: string
          description: Present when the failure is scoped to a single ticker.
      required:
        - error
        - code

    CandleSeries:
      type: object
      required: [ticker, interval, candles]
      properties:
        ticker:
          type: string
        interval:
          type: string
          enum: [5m, 30m, 1d]
        candles:
          type: array
          items:
            $ref: "#/components/schemas/Candle"

    AuthRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password

    AuthResponse:
      type: object
      properties:
        token:
          type: string
      required:
        - token

    ListItem:
      type: object
      properties:
        ticker:
          type: string
      required:
        - ticker

    List:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/ListItem"
      required:
        - id
        - name
        - items

    ListSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
      required:
        - id
        - name

    CreateListRequest:
      type: object
      properties:
        name:
          type: string
      required:
        - name

    RenameListRequest:
      type: object
      properties:
        name:
          type: string
      required:
        - name

    AddListItemRequest:
      type: object
      properties:
        ticker:
          type: string
      required:
        - ticker

    SentimentStreamEvent:
      oneOf:
        - type: object
          properties:
            type:
              type: string
              enum: [connected]
          required:
            - type
        - type: object
          description: Emitted for each newly scored article for a watched ticker.
          properties:
            type:
              type: string
              enum: [SOURCE_UPDATE]
            ticker:
              type: string
            source:
              $ref: "#/components/schemas/SourceResult"
            avgScore:
              type: number
              minimum: -1.0
              maximum: 1.0
              description: Recomputed average score for this ticker across all fresh source_score rows.
          required:
            - type
            - ticker
            - source
            - avgScore