# webclient/src/api/generated/dtos/sourceResult.gen.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Sentiment Search API
 * Three-stage pipeline: (1) stream stock info, (2) stream articles per ticker, (3) stream per-article sentiment scores.

 * OpenAPI spec version: 2.0.0
 */

export type SourceResult = {
  url: string;
  snippet: string;
  /**
     * Timestamp of the last update moment in seconds since Unix epoch
     * @minimum 0
     */
  updatedAtSec: number;
  /**
     * Timestamp of the scrape moment in seconds since Unix epoch
     * @minimum 0
     */
  scrapedAtSec: number;
} & {
  /**
     * @minimum -1
     * @maximum 1
     */
  score: number;
};