# webclient/src/api/generated/dtos/stock.gen.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Sentiment Search API
 * Three-stage pipeline: (1) stream stock info, (2) stream articles per ticker, (3) stream per-article sentiment scores.

 * OpenAPI spec version: 2.0.0
 */

export interface Stock {
  ticker: string;
  name: string;
}